trading_bot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
arrow==1.3.0
attrs==25.3.0
Authlib==1.5.2
backtesting==0.6.4
bitarray==3.4.1
bokeh==3.7.2
bravado==10.4.1
bravado-core==6.1.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==2.0.12
ckzg==2.1.1
click==8.2.0
contourpy==1.3.2
cryptography==45.0.2
cytoolz==1.0.1
dateparser==1.2.1
discord.py==2.5.2
distro==1.9.0
dotenv==0.9.9
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
fastapi==0.115.12
fqdn==1.5.1
frozenlist==1.6.0
greenlet==3.2.2
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
isoduration==20.11.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jwt==1.3.1
MarkupSafe==3.0.2
monotonic==1.6
msgpack==1.1.0
msgpack-python==0.5.6
multidict==6.4.4
narwhals==1.35.0
numpy==2.2.4
openai==1.79.0
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
passlib==1.7.4
pillow==11.2.1
propcache==0.3.1
pybit==5.11.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyMySQL==1.1.1
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyunormalize==16.0.0
PyYAML==6.0.2
redis==6.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rlp==4.1.0
rpds-py==0.25.1
schedule==1.2.2
setuptools==79.0.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
swagger-spec-validator==3.0.4
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
types-python-dateutil==2.9.0.20250516
types-requests==2.32.0.20250515
types-urllib3==1.26.25.14
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
ulid==1.1
ulid-py==1.1.0
uri-template==1.3.0
urllib3==2.4.0
uuid_utils==0.10.0
uvicorn==0.34.2
web3==7.12.0
webcolors==24.11.1
websocket-client==1.8.0
websockets==15.0.1
xyzservices==2025.1.0
yarl==1.20.0
